refactor: change parameter injection
