+++ src/main/java/com/organiza/api/modules/transactions/infra/database/entity/TransactionModel.java
