+++ src/main/java/com/organiza/api/modules/users/infra/database/entity/UserModel.java
