+++ src/main/java/com/organiza/api/modules/users/infra/database/repository/UserRepository.java
